+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Wyliodrin STUDIO 2.1.1-beta
   * **added** folder template for opcua model
   * **fixed** multiple instance popup appearing multiple times
+  * **fixed** gutter background color in electron
   * **update** electron to version 7
   * **update** serialport to version 8
   * **update** new color for opcua model folder