+++ CHANGELOG.md
@@ -2,6 +2,8 @@
   * **added** folder template for opcua model
   * **fixed** multiple instance popup appearing multiple times
   * **fixed** gutter background color in electron
+  * **fixed** image error from RPK simulator
+  * **fixed** display of extra brace at device os version
   * **update** electron to version 7
   * **update** serialport to version 8
   * **update** new color for opcua model folder